fix: route nostr helpers through client module
